LocalScanner: Generalize the filtering against non-license-filenames
A hard coded ignore list of files which should not be scanned was so far
applied only in the ScanCode integration. Other Scanner implementations
didn't adhere to that. For consistency, apply that ignore list to all
(local) scanners.

While using ScanCode's '--ignore' command line option might be a tiny
little bit better performance wise, stop using that for simplicity.
Apply the filtering only after storing the scan result in the scan
result storage in order to not invalidate the scan result storage
entries anymore after a future change to the ignore list.

Signed-off-by: Frank Viernau <frank.viernau@here.com>